refactor(tool): extract the byte bounds on tool output into ToolResultBounder
Tool results are untrusted bytes and cross two boundaries — the provider wire
(content) and the trace/inspector/persisted stream (artifacts). The UTF-8
coercion and the two independent byte caps now live in a stateless class of
their own, moved verbatim; the loop drops from 995 to 852 lines and the
"untrusted bytes" concern stops being interleaved with the control flow.

What deliberately did not move, per the seam analysis: everything else. The
three entry points share the gating resolution and the invoke seam, the two
resume refusal ladders differ in exactly the 15% that IS ADR-105, and the
suspension scans read five loop-local counters — every tempting extraction
there cuts the security path that ADR-093/108 deliberately concentrated.
invoke() remains the single seam that applies the bounds; the bounder owns
the how, never the where.

The bounder is a defaulted, NON-nullable constructor argument — the
JsonSchemaValidator pattern ADR-120 chose over nullability — so no wiring can
leave the caps absent, and all nine positional and named test constructions
compile unchanged. Slot order 1-10 is byte-identical.

The build itself proved the regression net works: the extracted class briefly
lacked the JsonException import, the catch silently never fired, and
unencodableArtifactCollapsesToTheFailClosedMarker failed immediately.

Zero test-body changes; the six capping tests exercise the moved code
end-to-end through runLoop exactly as before.
